requirements.txt
opencv-python>=4.5.0
numpy>=1.20.0
Pillow>=8.0.0
pandas>=1.3.0
